LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/report-to-directive-allowed-in-meta.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Report is observable to ReportingObserver assert_equals: expected 54 but got 0
PASS Violation report status OK.
